Extract writer for outputting log messages
