+++ theme.py
@@ -1,0 +1,31 @@
+TAG_BLACK = 'black'
+TAG_BLUE = 'blue'
+TAG_GREEN = 'green'
+TAG_GREY = 'grey'
+TAG_ORANGE = 'orange'
+TAG_RED = 'red'
+
+THEME_DEFAULT = 'default'
+THEME_DARK = 'dark'
+THEME_LIGHT = 'light'
+
+theme_config = {
+    THEME_DARK: {
+        TAG_BLACK: 'light grey',
+        TAG_BLUE: '#7289DA',
+        TAG_GREEN: '#4E9D4E',
+        TAG_GREY: 'grey',
+        TAG_ORANGE: '#FF981F',
+        TAG_RED: '#F04747',
+        'background': '#36393E'
+    },
+    THEME_LIGHT: {
+        TAG_BLACK: 'black',
+        TAG_BLUE: 'blue',
+        TAG_GREEN: 'green',
+        TAG_GREY: 'grey',
+        TAG_ORANGE: 'orange',
+        TAG_RED: 'red',
+        'background': 'white'
+    }
+}
